- changed names, init, fixed directories.py, moved models.py
