+++ indexer/src/security.py
@@ -3,7 +3,7 @@ from fastapi import Request
 from .settings import settings
 
 
-def checkToken(request: Request) -> bool:
+def check_token(request: Request) -> bool:
     if os.path.basename(request.url.path) in settings.private_paths:
-        return request.headers.get("Token", None) == settings.token
+        return request.headers.get("Token") == settings.token
     return True
